refactor(calendar): dedupe repostStatus resolution in EventService
The SharedEventEntity-first / EventRepostEntity-fallback repostStatus
resolution loop was copied across three EventService methods
(listEvents, bulkAssignCategories, replaceEventCategories). Extract the
shared derivation into a private resolveRepostStatusForCalendar helper
that returns a Map<eventId, 'auto' | 'manual'>; call sites apply the
'none' default via `map.get(id) ?? 'none'`, preserving identical
precedence and resulting repostStatus values.

Also consolidate the duplicated buildMockApInterface test stub into a
shared helper at test/EventService/helpers/ap-interface.ts (full 3-arg
signature) and import it from the three EventService test files that
carried their own copies.
